Create README.md
This README is for the Hijrah calendar. I see that nowadays many people are forgetting about this calendar system. In this repository, the Hijrah calendar is the main calendar while the Masihi calendar is the secondary (alternative) calendar. The dates in this calendar are one day faster than the hanging calendars in homes because this Hijrah calendar follows "tiba-tiba Hari Raya" and "raya kelam-kabut" that happened on Syawal 1443H.